Fix crash from not spawning Market

+++ game_common/src/economy/households/bakery/mod.rs
@@ -151,7 +151,7 @@ impl Sleeper for Bakery {
 }
 
 impl EvaluationRequester for Bakery {
-    fn expect_n_results(&mut self, _r: Resource, _n: usize, _: &mut World) {}
+    fn expect_n_results(&mut self, _r: Resource, _n: u32, _: &mut World) {}
     fn on_result(&mut self, _e: &EvaluatedSearchResult, _: &mut World) {}
 }
