+++ game_common/src/economy/households/grain_farm/mod.rs
@@ -124,7 +124,7 @@ impl Sleeper for GrainFarm {
 }
 
 impl EvaluationRequester for GrainFarm {
-    fn expect_n_results(&mut self, _r: Resource, _n: usize, _: &mut World) {}
+    fn expect_n_results(&mut self, _r: Resource, _n: u32, _: &mut World) {}
     fn on_result(&mut self, _e: &EvaluatedSearchResult, _: &mut World) {}
 }
