+++ game_common/src/economy/households/grocery_shop/mod.rs
@@ -150,7 +150,7 @@ impl Household for GroceryShop {
 use super::ResultAspect;
 
 impl EvaluationRequester for GroceryShop {
-    fn expect_n_results(&mut self, resource: Resource, n: usize, world: &mut World) {
+    fn expect_n_results(&mut self, resource: Resource, n: u32, world: &mut World) {
         self.update_results(resource, &ResultAspect::SetTarget(n), world);
     }
